refactor: DiscountFactory -> DiscountType 클래스명 수정
